+++ changelog.md
@@ -4,6 +4,7 @@
 
 - **Fixed:**
   - Auto update: when error happens correctly re-enable menu item
+  - Code editor: Add bracket colorization matching
 - **Changed:**
   - Update electron 13.3.0 to 13.5.0
   - Update fswin 2.21.905 to 2.21.929